remove unused service item table
