+++ src/components/Office/ServiceItemDetails/ServiceItemDetails.module.scss
@@ -15,7 +15,7 @@
   @include u-margin-top(2);
 }
 
-.detailImage {
+.detailCrating {
   @include u-display(flex);
   p {
     @include u-margin-top(0);
